% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{cal_model_accuracy}
\alias{cal_model_accuracy}
\title{Title}
\usage{
cal_model_accuracy(model, newset, keycol, calf, caltype)
}
\arguments{
\item{model}{}

\item{newset}{}

\item{keycol}{}

\item{calf}{}

\item{caltype}{}
}
\description{
Title
}
