% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{collect_test_set}
\alias{collect_test_set}
\title{Title}
\usage{
collect_test_set(name_para_sets, factordf = 1, renew = TRUE)
}
\arguments{
\item{name_para_sets}{}

\item{factordf}{}

\item{renew}{}
}
\description{
Title
}
