% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_advbase_set}
\alias{get_advbase_set}
\title{Title}
\usage{
get_advbase_set(umid_sql, neg_umid_sql, specDate, negspecDate = specDate,
  renew = FALSE, complexity = FALSE, sign = 1)
}
\arguments{
\item{umid_sql}{}

\item{neg_umid_sql}{}

\item{specDate}{}

\item{negspecDate}{}

\item{renew}{}

\item{complexity}{}

\item{sign}{}
}
\description{
Title
}
