% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_base_with_restrict}
\alias{get_base_with_restrict}
\title{Title}
\usage{
get_base_with_restrict(umid_sql, specDate = spec.date.str, mode = "base",
  mobile = TRUE)
}
\arguments{
\item{umid_sql}{}

\item{specDate}{}

\item{mode}{}

\item{mobile}{}
}
\description{
Title
}
