% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_base_with_restrict_union}
\alias{get_base_with_restrict_union}
\title{Title}
\usage{
get_base_with_restrict_union(umid_sql_list, datelist, mode = "base",
  mobile = TRUE)
}
\arguments{
\item{umid_sql_list}{}

\item{datelist}{}

\item{mode}{}

\item{mobile}{}
}
\description{
Title
}
