% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_complex_train_set}
\alias{get_complex_train_set}
\title{Title}
\usage{
get_complex_train_set(dateset, umid_sql, renew = FALSE)
}
\arguments{
\item{dateset}{}

\item{umid_sql}{}

\item{renew}{}
}
\description{
Title
}
