% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_test_sample}
\alias{get_test_sample}
\title{Title}
\usage{
get_test_sample(city, specDate, func_umid_sql = 1, factordf = 1,
  renew = FALSE)
}
\arguments{
\item{city}{}

\item{specDate}{}

\item{func_umid_sql}{}

\item{factordf}{}

\item{renew}{}
}
\description{
Title
}
