% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_test_sample_result}
\alias{get_test_sample_result}
\title{Title}
\usage{
get_test_sample_result(city, specDate, func_umid_sql = 1, factordf = 1,
  test_model)
}
\arguments{
\item{city}{}

\item{specDate}{}

\item{func_umid_sql}{}

\item{factordf}{}

\item{test_model}{}
}
\description{
Title
}
