% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_var_sample}
\alias{get_var_sample}
\title{Title}
\usage{
get_var_sample(oringin_set, keycol = "sign", posdiv = 0.2, negdiv = 1,
  randnum = 999)
}
\arguments{
\item{oringin_set}{}

\item{keycol}{}

\item{posdiv}{}

\item{negdiv}{}

\item{randnum}{}
}
\description{
Title
}
