% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{pred_day_result}
\alias{pred_day_result}
\title{Title}
\usage{
pred_day_result(training_model, specDate = last.date.str, renew = TRUE)
}
\arguments{
\item{training_model}{}

\item{specDate}{}

\item{renew}{}
}
\description{
Title
}
