% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{run_flexible_sql}
\alias{run_flexible_sql}
\title{Title}
\usage{
run_flexible_sql(SQL, specDate = last.date.str, specCity = city_name)
}
\arguments{
\item{SQL}{}

\item{specDate}{}

\item{specCity}{}
}
\description{
Title
}
